fix(tests) jest mock on BuiltInDependencies
